+++ scripts/build.sh
@@ -6,3 +6,5 @@ echo build historical
 cargo build --no-default-features --features historical
 echo build live
 cargo build --no-default-features --features live
+echo build examples
+cargo build --examples